Show/Hide Login modal
Toggling modal visibility using this.$emit and props.
Header component and Modal components emit methods which are defined in App.vue. There based on the payload, a state is being updated which is then passed to Modal component via a prop. A v-if finally decides whether or not to display the login modal using the prop value.